# ENH: sparse: Speed up `_add_sparse` for DIA format