This PR rewrites the `_add_sparse` logic for DIA-format sparse arrays/matrices to avoid expensive allocations and Python looping.

My local benchmarks show that this is faster, in the easy case where the sparsity pattern doesn't change, but also in the more complex case where the sets of diagonal offsets are non-overlapping.

![no_overlap](https://github.com/scipy/scipy/assets/82855/82c0a004-b15f-438d-9c62-77062f56eb6c)
![trivial](https://github.com/scipy/scipy/assets/82855/0d077393-7eb7-417b-9d01-6ff3bc97511f)

Generated via: 
[bench.py](https://github.com/scipy/scipy/files/15330327/bench.py.txt)