@perimosocordiae 2024-05-16
I ran the `asv` benchmarks, though notably I had to remove the `--quick` flag from our dev.py runner first. If running in `--quick` mode, the results are too noisy to be useful.

| Change   | Before [8ce393f1] <main>   | After [8eb26148] <dia-add>   |   Ratio | Benchmark (Parameter)                                     |
|----------|----------------------------|------------------------------|---------|-----------------------------------------------------------|
| -        | 811±50μs                   | 265±40μs                     |    0.33 | sparse.Arithmetic.time_arithmetic('dia', 'AA', '__add__') |

In these benchmarks the A matrix is in DIA format, and B is in CSR format. So the AB, BA, and BB tests aren't exercising the updated code path, which is why they don't show up in the timing diffs.