@perimosocordiae 2024-05-18
Thanks for pointing out the opportunity for the other direction's overlapping-offsets fast path, @dschult. I ran a before and after using a similar setup to what you described, where `c`'s offsets are a subset of `a`'s offsets:

Before 1eba3cb: 
![before](https://github.com/scipy/scipy/assets/82855/c8de6f38-5557-43c6-a4bc-5d84b3cbb7c7)

After 1eba3cb:
![after](https://github.com/scipy/scipy/assets/82855/188cbf43-fc82-43a9-a604-13c5c8057e15)